[state-sync] Move networking into consensus
To make networking as effective as possible, it makes sense to host the
interface into networking in the code itself. This allows for binding
the types sent across the network into local types. While this isn't
done in this PR, this is part of the future goals. This PR is just a
bridge to that future.

At the same time, this PR removes the verbose and unnecessary testing of
the underlying network connectivity that has already been established
upstream. There may be some benefit in making a suite to verify these
rather trivial interfaces in the future, but it adds unnecessary
overhead in code management at this stage due to its negligible
contribution to additional testing.
