+++ network/src/counters.rs
@@ -94,16 +94,6 @@ pub static PENDING_NETWORK_REQUESTS: Lazy<IntCounterVec> = Lazy::new(|| {
     .unwrap()
 });
 
-/// Counter of pending network events to Consensus
-pub static PENDING_CONSENSUS_NETWORK_EVENTS: Lazy<IntCounterVec> = Lazy::new(|| {
-    register_int_counter_vec!(
-        "pending_consensus_network_events",
-        "Counters(queued,dequeued,dropped) related to pending network notifications to Consensus",
-        &["state"]
-    )
-    .unwrap()
-});
-
 /// Counter of pending network events to Health Checker.
 pub static PENDING_HEALTH_CHECKER_NETWORK_EVENTS: Lazy<IntCounterVec> = Lazy::new(|| {
     register_int_counter_vec!(
